admin/requirements.txt
firebase-admin==5.0.1
Pillow==8.2.0
ipython==7.24.1
